Fast query of WITCH results files

Returns a data.table containing the result of the query with additional columns on scenario

witch_query(
  item = NULL,
  resgdx = NULL,
  filter = list(),
  add_scenario = TRUE,
  scenario_mapping = witch_scen_name,
  add_year = TRUE,
  year_mapping = witch_period_year,
  valigdx = NULL,
  histgdx = NULL,
  agg_n = NULL,
  clean_columns = TRUE,
  ...
)

Arguments

item

parameter or variable name

resgdx

list of results gdx from WITCH

filter

named list of queries

add_year

convert t into year

year_mapping

a mapping table to translate t into year

add_scen

convert gdx into scenario name

scen_table

a conversion function or a mapping table to translate gdx into scenario name

See also